@@ -3,27 +3,38 @@ import 'package:overtimer_mobile/models/interval/interval_item.dart';
 import 'package:http/http.dart' as http;
 import 'dart:convert' as convert;
 
+import 'package:overtimer_mobile/models/tag_item.dart';
 import 'auth_service.dart';
 
 class IntervalService {
-  // static Future<void> deleteTag(int id) async {
-  //   try {
-  //     var apiUrl = dotenv.get("API_URL");
-  //     var url = Uri.http(apiUrl, '/api/Tag/$id');
+  static Future<void> deleteInterval(int id) async {
+    try {
+      var apiUrl = dotenv.get("API_URL");
+      var url = Uri.http(apiUrl, '/api/Work/$id');
 
-  //     var response = await http.delete(url);
+      var response = await http.delete(url);
 
-  //     if (response.statusCode == 200) {
-  //     } else {
-  //       throw Exception(
-  //           'Status code: ${response.statusCode}, Body: ${response.body}');
-  //     }
-  //   } catch (e) {
-  //     throw Exception(e);
-  //   }
-  // }
+      if (response.statusCode == 200) {
+      } else {
+        throw Exception(
+            'Status code: ${response.statusCode}, Body: ${response.body}');
+      }
+    } catch (e) {
+      throw Exception(e);
+    }
+  }
+
   DateTime _createDateTime(String datetime) {
     return DateTime.parse(datetime);
+  }
+
+  List<TagItem> _getTagIds(List intervalTags) {
+    return intervalTags.map((tag) {
+      return TagItem(
+          id: tag['id'] as int,
+          name: tag['name'] as String,
+          companyId: tag['companyId'] as int);
+    }).toList();
   }
 
   Future<List<IntervalItem>> getIntervals() async {
@@ -44,16 +55,16 @@ class IntervalService {
         List<dynamic> decodedData = convert.jsonDecode(response.body);
 
         List<IntervalItem> intervals = decodedData
-            .map((dynamic jsonTag) => IntervalItem(
-                  id: jsonTag['id'] as int? ?? 0,
-                  title: jsonTag['title'] as String? ?? "",
-                  // start: jsonTag['initialDateTime'] as DateTime? ?? 0,
-                  // end: jsonTag['finalDateTime'] as DateTime? ?? 0,
-                  start: _createDateTime(jsonTag['initialDateTime']),
-                  end: _createDateTime(jsonTag['finalDateTime']),
+            .map((dynamic jsonInterval) => IntervalItem(
+                  id: jsonInterval['id'] as int? ?? 0,
+                  userId: jsonInterval['userId'] as int? ?? 0,
+                  projectId: jsonInterval['companyId'] as int? ?? 0,
+                  tagIds: _getTagIds(jsonInterval['tags']),
+                  title: jsonInterval['title'] as String? ?? "",
+                  start: _createDateTime(jsonInterval['initialDateTime']),
+                  end: _createDateTime(jsonInterval['finalDateTime']),
                 ))
             .toList();
-
         return intervals;
       } else {
         throw Exception('Falha ao carregar os dados: ${response.body}');
@@ -101,10 +112,12 @@ class IntervalService {
       };
 
       var requestBody = {
-        'companyId': companyId,
+        'userId': interval.userId,
+        'projectId': interval.projectId,
         'title': interval.title,
-        'initialDateTime': interval.start,
-        'endDateTime': interval.end,
+        'initialDateTime': interval.start.toUtc().toIso8601String(),
+        'finalDateTime': interval.end.toUtc().toIso8601String(),
+        'tagIds': interval.tagIds.map((tag) => tag.id).toList(),
       };
 
       var response = await http.post(
@@ -122,31 +135,36 @@ class IntervalService {
     }
   }
 
-  // static Future<void> editTag(int id, String name, int companyId) async {
-  //   try {
-  //     var apiUrl = dotenv.get("API_URL");
-  //     var url = Uri.http(apiUrl, '/api/Tag/$id');
+  static Future<void> editInterval(IntervalItem interval, int companyId) async {
+    try {
+      var apiUrl = dotenv.get("API_URL");
+      var url = Uri.http(apiUrl, '/api/Work/${interval.id}');
 
-  //     var headers = {'Content-Type': 'application/json'};
+      var headers = {'Content-Type': 'application/json'};
 
-  //     var requestBody = {
-  //       'companyId': companyId,
-  //       'name': name,
-  //     };
+      var requestBody = {
+        'userId': interval.userId,
+        'projectId': interval.projectId,
+        'title': interval.title,
+        'initialDateTime': interval.start.toUtc().toIso8601String(),
+        'finalDateTime': interval.end.toUtc().toIso8601String(),
+        'tagIds': interval.tagIds.map((tag) => tag.id).toList(),
+      };
 
-  //     var response = await http.patch(
-  //       url,
-  //       headers: headers,
-  //       body: convert.jsonEncode(requestBody),
-  //     );
+      var response = await http.patch(
+        url,
+        headers: headers,
+        body: convert.jsonEncode(requestBody),
+      );
 
-  //     if (response.statusCode == 200) {
-  //       // Tag editada com sucesso
-  //     } else {
-  //       throw Exception('Erro na edição da tag: ${response.statusCode}');
-  //     }
-  //   } catch (e) {
-  //     throw Exception('Exception during HTTP request: $e');
-  //   }
-  // }
+      if (response.statusCode == 200) {
+        // Intervalo editado com sucesso
+      } else {
+        throw Exception('Erro na edição da tag: ${response.statusCode}');
+      }
+    } catch (e) {
+      print(e);
+      throw Exception('Exception during HTTP request: $e');
+    }
+  }
 }
