@@ -1,9 +1,11 @@
 import 'package:flutter/material.dart';
 
 class NewIntervalTitleInput extends StatefulWidget {
-  const NewIntervalTitleInput({super.key, required this.onChange});
+  const NewIntervalTitleInput(
+      {super.key, this.initialValue = '', required this.onChange});
 
   final void Function(String newValue) onChange;
+  final String initialValue;
 
   @override
   State<NewIntervalTitleInput> createState() => _NewIntervalTitleInputState();
@@ -15,6 +17,7 @@ class _NewIntervalTitleInputState extends State<NewIntervalTitleInput> {
   @override
   void initState() {
     super.initState();
+    textController.text = widget.initialValue;
     textController.addListener(() {
       widget.onChange(textController.text);
     });
